requirements.txt
scrapy>=2.5.0
pymongo>=4.0.0
python-slugify>=6.0.0
itemadapter>=0.7.0
requests>=2.28.0